% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dom.freq.R
\name{dom.freq}
\alias{dom.freq}
\title{dom.freq}
\usage{
dom.freq()
}
\arguments{
\item{dom.freq}{Estimates the dominant frequency, using the function in seewave v.2.1.0 package)}
}
\value{
return a table with a different soundecology index
}
\description{
dom.freq
}
\examples{
#firt, created a vector call: "time" and define your sequence time
#time<-seq(0:3)
#dom.freq()->Fdom.freq #The vector "Fdom.freq" call your result
#Fdom.freq #Call the vector, and now you can see all index by channel.


}
\author{
Oscar Ramirez Alan (\email{osoramirez@gmail.com})
}
