% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dom.freq6.R
\name{dom.freq6}
\alias{dom.freq6}
\title{dom.freq6}
\usage{
dom.freq6()
}
\arguments{
\item{dom.freq6}{Estimates the dominant frequency, using the function in seewave package)}
}
\value{
return a table with a different soundecology index
}
\description{
dom.freq6
}
\examples{
#This code analyzes for a 6-minute interval
#dom.freq6()->dom.freq6min #Vector that contain your results
#dom.freq6min #Call the vector, and see your results.

}
\references{
{
Luis J. Villanueva-Rivera and Bryan C. Pijanowski (2018). soundecology: Soundscape Ecology. R package version 1.3.3. https://CRAN.R-project.org/package=soundecology.

Sueur, J., Aubin, T., & Simonis, C. (2008). Seewave, a free modular tool for sound analysis and synthesis. Bioacoustics, 18(2), 213-226.

Uwe Ligges, Sebastian Krey, Olaf Mersmann, and Sarah Schnackenberg (2018). tuneR: Analysis of Music and Speech. URL: https://CRAN.R-project.org/package=tuneR.
}
}
\author{
Oscar Ramírez Alán (\email{osoramirez@gmail.com}). Implements a
loops using base function from seewave and soundecology.
}
