% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{evaluation}
\alias{evaluation}
\title{evaluation!}
\usage{
evaluation()
}
\arguments{
\item{evaluation}{Estimates a evaluation of insect and rain}
}
\value{
a evaluation of rain and insects
}
\description{
evaluation!
}
\examples{
#firt, created a vector call: "time" and define your sequence time
#time<-seq(0:3)
#evaluation()->evaluation #Remember always chance the vector name
#evaluation
}
