% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation10.R
\name{evaluation10}
\alias{evaluation10}
\title{Estimate if your file contain rain and insects for 10 min intervals}
\usage{
evaluation10()
}
\arguments{
\item{evaluation10}{Estimates a evaluation of insect and rain}
}
\value{
a evaluation of rain and insects
}
\description{
Estimate if your file contain rain and insects for 10 min intervals
}
\examples{
#This code analyzes for a 10-minute interval
#evaluation10()->evaluation10min #Vector that contain your results
#evaluation10min #Call the vector, and see your results.



}
\references{
{
Luis J. Villanueva-Rivera and Bryan C. Pijanowski (2018). soundecology: Soundscape Ecology. R package version 1.3.3. https://CRAN.R-project.org/package=soundecology.

Sueur, J., Aubin, T., & Simonis, C. (2008). Seewave, a free modular tool for sound analysis and synthesis. Bioacoustics, 18(2), 213-226.

Uwe Ligges, Sebastian Krey, Olaf Mersmann, and Sarah Schnackenberg (2018). tuneR: Analysis of Music and Speech. URL: https://CRAN.R-project.org/package=tuneR.
}
}
\author{
Oscar Ramírez Alán(\email{osoramirez@gmail.com}). Implements a
loops using base function from seewave and soundecology.
}
