% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerindex.R
\name{powerindex}
\alias{powerindex}
\title{powerindex}
\usage{
powerindex()
}
\arguments{
\item{powerindex}{Estimates the power spectral analysis of your sound file in dB by frequency of 1kHz.}
}
\value{
return a table with a different power spectral index
}
\description{
powerindex
}
\examples{
#firt, created a vector call: "time" and define your sequence time
#time<-seq(0:3)
#powerindex()->powervector #Remember always chance the vector name
#powervector #Call the vector, and now you can see all index by channel.

}
\author{
Oscar Ramirez Alan (\email{osoramirez@gmail.com})
}
