% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerindex1.R
\name{powerindex1}
\alias{powerindex1}
\title{Estimate power spectral density (PSD) in 1 min intervals}
\usage{
powerindex1()
}
\arguments{
\item{powerindex1}{Estimates the power spectral analysis of your sound file in dB by frequency of 1kHz.}
}
\value{
return a table with a different power spectral index
}
\description{
Estimate power spectral density (PSD) in 1 min intervals
}
\examples{
#This code analyzes for a 1-minute interval
#powerindex1()->powerindex1min #Vector that contain your results
#powerindex1min #Call the vector, and see your results.

}
\references{
{
Luis J. Villanueva-Rivera and Bryan C. Pijanowski (2018). soundecology: Soundscape Ecology. R package version 1.3.3. https://CRAN.R-project.org/package=soundecology.

Sueur, J., Aubin, T., & Simonis, C. (2008). Seewave, a free modular tool for sound analysis and synthesis. Bioacoustics, 18(2), 213-226.

Uwe Ligges, Sebastian Krey, Olaf Mersmann, and Sarah Schnackenberg (2018). tuneR: Analysis of Music and Speech. URL: https://CRAN.R-project.org/package=tuneR.
}
}
\author{
Oscar Ramírez Alán (\email{osoramirez@gmail.com}). Implements a
loops using base function from seewave and soundecology.
}
