% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerindex10.R
\name{powerindex10}
\alias{powerindex10}
\title{Estimate power spectral density (PSD) in 10 min intervals}
\usage{
powerindex10()
}
\arguments{
\item{powerindex10}{Estimates the power spectral analysis of your sound file in dB by frequency of 1kHz.}
}
\value{
return a table with a different power spectral index
}
\description{
Estimate power spectral density (PSD) in 10 min intervals
}
\examples{
#This code analyzes for a 10-minutes intervals
#powerindex10()->powerindex10min #Vector that contain your results
#powerindex10min #Call the vector, and see your results.

}
\references{
{
Luis J. Villanueva-Rivera and Bryan C. Pijanowski (2018). soundecology: Soundscape Ecology. R package version 1.3.3. https://CRAN.R-project.org/package=soundecology.

Sueur, J., Aubin, T., & Simonis, C. (2008). Seewave, a free modular tool for sound analysis and synthesis. Bioacoustics, 18(2), 213-226.

Uwe Ligges, Sebastian Krey, Olaf Mersmann, and Sarah Schnackenberg (2018). tuneR: Analysis of Music and Speech. URL: https://CRAN.R-project.org/package=tuneR.
}
}
\author{
Oscar Ramírez Alán (\email{osoramirez@gmail.com}). Implements a
loops using base function from seewave and soundecology.
}
