% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundindex.R
\name{soundindex}
\alias{soundindex}
\title{soundindex}
\usage{
soundindex()
}
\arguments{
\item{soundindex}{Estimates sound index analysis. This function using a loop that include the most common sound ecology index from (soundecolgy v. 1.3.3 and seewave v.2.1.0 package)}
}
\value{
return a table with a different soundecology index
}
\description{
soundindex
}
\examples{
#firt, created a vector call: "time" and define your sequence time
#time<-seq(0:3)
#soundindex()->filevector #Remember always chance the vector name
#filevector #Call the vector, and now you can see all index by channel.


}
\author{
Oscar Ramirez Alan (\email{osoramirez@gmail.com})
}
