% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundindex60.R
\name{soundindex60}
\alias{soundindex60}
\title{Estimates sound index analysis for 60 min intervals}
\usage{
soundindex60()
}
\arguments{
\item{soundindex60}{Estimates sound index analysis. This function using a loop that include the most common sound ecology index from (soundecolgy v. 1.3.3 and seewave v.2.1.0 package)}
}
\value{
return a table with a different soundecology index
}
\description{
Estimates sound index analysis for 60 min intervals
}
\examples{
#This code analyzes for a 60-minutes intervals
#soundindex60()->soundindex60min #Vector that contain your results
#soundindex60min #Call the vector, and see your results.

}
\references{
{
Luis J. Villanueva-Rivera and Bryan C. Pijanowski (2018). soundecology: Soundscape Ecology. R package version 1.3.3. https://CRAN.R-project.org/package=soundecology.

Sueur, J., Aubin, T., & Simonis, C. (2008). Seewave, a free modular tool for sound analysis and synthesis. Bioacoustics, 18(2), 213-226.

Uwe Ligges, Sebastian Krey, Olaf Mersmann, and Sarah Schnackenberg (2018). tuneR: Analysis of Music and Speech. URL: https://CRAN.R-project.org/package=tuneR.
}
}
\author{
Oscar Ramírez Alán (\email{osoramirez@gmail.com}). Implements a
loops using base function from seewave and soundecology.
}
