% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spl.R
\name{spl}
\alias{spl}
\title{Sound pressure level}
\usage{
spl
}
\arguments{
\item{spl}{Sound pressure level uses a logarithmic scale to represent the sound pressure of a sound relative to a reference pressure}
}
\value{
Sound presure level
}
\description{
Sound pressure level
}
\examples{
#firt, created a vector call: "time" and define your sequence time
#time<-seq(0:3)
#spl()->spl1 #Remember always chance the vector name
#spl1

}
