% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spl1.R
\name{spl1}
\alias{spl1}
\title{Sound pressure level at 1 min interval}
\usage{
spl1
}
\arguments{
\item{spl1}{Sound pressure level uses a logarithmic scale to represent the sound pressure of a sound relative to a reference pressure}
}
\value{
Sound presure level
}
\description{
Sound pressure level at 1 min interval
}
\examples{
#This code analyzes for a 1-minute interval
#spl1()->spl1min #Vector that contain your results
#spl1min #Call the vector, and see your results.

}
\references{
{
Uwe Ligges, Sebastian Krey, Olaf Mersmann, and Sarah Schnackenberg (2018). tuneR: Analysis of Music and Speech. URL: https://CRAN.R-project.org/package=tuneR
}
}
\author{
Oscar Ramírez (\email{osoramirez@gmail.com}). Implements a
loops using base function from seewave and soundecology.
}
