% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spl10.R
\name{spl10}
\alias{spl10}
\title{Sound pressure level at 10 min interval}
\usage{
spl10
}
\arguments{
\item{spl10}{Sound pressure level uses a logarithmic scale to represent the sound pressure of a sound relative to a reference pressure}
}
\value{
Sound presure level
}
\description{
Sound pressure level at 10 min interval
}
\examples{
#This code analyzes for a 10-minute interval
#spl10()->spl10min #Vector that contain your results
#spl10min #Call the vector, and see your results.

}
\references{
{
Uwe Ligges, Sebastian Krey, Olaf Mersmann, and Sarah Schnackenberg (2018). tuneR: Analysis of Music and Speech. URL: https://CRAN.R-project.org/package=tuneR.
}
}
\author{
Oscar Ramírez (\email{osoramirez@gmail.com}). Implements a
loops using base function from seewave and soundecology.
}
