% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spl6.R
\name{spl6}
\alias{spl6}
\title{Sound pressure level at 6 min interval}
\usage{
spl6
}
\arguments{
\item{spl6}{Sound pressure level uses a logarithmic scale to represent the sound pressure of a sound relative to a reference pressure}
}
\value{
Sound presure level
}
\description{
Sound pressure level at 6 min interval
}
\examples{
#This code analyzes for a 1-minute interval
#spl6()->spl6min #Remember always chance the vector name
#spl6min #Call the vector, and see your results.

}
\references{
{
Uwe Ligges, Sebastian Krey, Olaf Mersmann, and Sarah Schnackenberg (2018). tuneR: Analysis of Music and Speech. URL: https://CRAN.R-project.org/package=tuneR
}
}
\author{
Oscar Ramírez (\email{osoramirez@gmail.com}). Implements a
loops using base function from seewave and soundecology.
}
