% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spl60.R
\name{spl60}
\alias{spl60}
\title{Soundpressure level at 60 min interval}
\usage{
spl60
}
\arguments{
\item{spl60}{Sound pressure level uses a logarithmic scale to represent the sound pressure of a sound relative to a reference pressure}
}
\value{
Sound presure level
}
\description{
Soundpressure level at 60 min interval
}
\examples{
#This code analyzes for a 1-minute interval
#spl60()->spl60min #Vector that contain your results
#spl60min #Call the vector, and see your results.
}
