% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPL.R
\name{SPL}
\alias{SPL}
\title{SPL!}
\usage{
SPL
}
\arguments{
\item{SPL}{Sound pressure level uses a logarithmic scale to represent the sound pressure of a sound relative to a reference pressure}
}
\value{
Sound presure level
}
\description{
SPL!
}
\examples{
#firt, created a vector call: "time" and define your sequence time
#time<-seq(0:3)
#SPL()->SPL #Remember always chance the vector name
#SPL
}
